Added timesteps and noise_schedule arguments (#91)
I have added the noise_schedule and timesteps arguments to the ElucidatedImagen class, so you can choose them when creating the model. If you think it is not necessary and that only the default should be there, reject the pull request.

Thanks again for the quick implementation! I will give it a try and report results 🚀